split linked-lists, bitmaps and storage-points from libnetdata.h
